# Split stage 2 in SAX fashion