--- a/src/generic/stage2/logger.h
+++ b/src/generic/stage2/logger.h
@@ -28,8 +28,8 @@ namespace logger {
     if (LOG_ENABLED) {
       log_depth = 0;
       printf("\n");
-      printf("| %-*s | %-*s | %-*s | %-*s | %-*s | Detail |\n", LOG_EVENT_LEN, "Event", LOG_BUFFER_LEN, "Buffer", LOG_SMALL_BUFFER_LEN, "Next", 5, "Next#", 5, "Tape#");
-      printf("|%.*s|%.*s|%.*s|%.*s|%.*s|--------|\n", LOG_EVENT_LEN+2, DASHES, LOG_BUFFER_LEN+2, DASHES, LOG_SMALL_BUFFER_LEN+2, DASHES, 5+2, DASHES, 5+2, DASHES);
+      printf("| %-*s | %-*s | %-*s | %-*s | Detail |\n", LOG_EVENT_LEN, "Event", LOG_BUFFER_LEN, "Buffer", LOG_SMALL_BUFFER_LEN, "Next", 5, "Next#");
+      printf("|%.*s|%.*s|%.*s|%.*s|--------|\n", LOG_EVENT_LEN+2, DASHES, LOG_BUFFER_LEN+2, DASHES, LOG_SMALL_BUFFER_LEN+2, DASHES, 5+2, DASHES);
     }
   }
 
@@ -71,7 +71,7 @@ namespace logger {
       } else {
         printf("| %-*s ", LOG_INDEX_LEN, "");
       }
-      printf("| %*u ", LOG_INDEX_LEN, structurals.next_tape_index());
+      // printf("| %*u ", LOG_INDEX_LEN, structurals.next_tape_index());
       printf("| %-s ", detail);
       printf("|\n");
     }