--- a/src/generic/stage2/structural_iterator.h
+++ b/src/generic/stage2/structural_iterator.h
@@ -6,13 +6,13 @@ class structural_iterator {
 public:
   const uint8_t* const buf;
   uint32_t *next_structural;
-  dom_parser_implementation &parser;
+  dom_parser_implementation &dom_parser;
 
   // Start a structural 
-  really_inline structural_iterator(dom_parser_implementation &_parser, size_t start_structural_index)
-    : buf{_parser.buf},
-      next_structural{&_parser.structural_indexes[start_structural_index]},
-      parser{_parser} {
+  really_inline structural_iterator(dom_parser_implementation &_dom_parser, size_t start_structural_index)
+    : buf{_dom_parser.buf},
+      next_structural{&_dom_parser.structural_indexes[start_structural_index]},
+      dom_parser{_dom_parser} {
   }
   // Get the buffer position of the current structural character
   really_inline const uint8_t* current() {
@@ -36,14 +36,14 @@ public:
     return buf[*(next_structural++)];
   }
   really_inline size_t remaining_len() {
-    return parser.len - *(next_structural-1);
+    return dom_parser.len - *(next_structural-1);
   }
 
   really_inline bool at_end() {
-    return next_structural == &parser.structural_indexes[parser.n_structural_indexes];
+    return next_structural == &dom_parser.structural_indexes[dom_parser.n_structural_indexes];
   }
   really_inline bool at_beginning() {
-    return next_structural == parser.structural_indexes.get();
+    return next_structural == dom_parser.structural_indexes.get();
   }
 };