… src/generic/stage2/tape_builder.h
… src/haswell/dom_parser_implementation.cpp
… src/westmere/dom_parser_implementation.cpp