This PR splits stage 2 in half to create a (non-public) SAX interface. The purpose of starting with stage 2 is to ensure we start in a good place performance-wise, rather than building something separate and struggling to match the performance we already have in stage 2. It has a fairly clean separation of concerns, with structural_parser in particular never touching the output document.

* structural_parser handles JSON iteration and array / object / colon / comma matching. It maintains depth, structural_index iteration and is_array[] state. It does not reference or touch the document, tape, string_buf, or count.
* tape_builder handles tape construction and primitive parsing. It maintains tape index, string_buf index, and count. It is passed uint8_t *'s into the buffer, and never references structural indexes or is_array[] state.

structural_parser::parse() is templatized on the tape_builder type, and calls into it with SAX-ish functions such as start_array, end_object, and parse_primitive. These are not final public names; the intent here is to get everything into the right *shape* before committing to a final interface.

The hope here is we can build alternate DOMs by hooking up new tape_builder implementations, as well as use the interface in language bindings, to avoid the intermediate serialization to tape and go straight to the language's types. Additionally, there should be opportunities to write raw unparsed DOMs, JSON pointer, and other things between.

It does some restructuring as well, which both simplify stage 2 and gain back the performance that would be lost (and then some) from a purely naive translation into SAX.

## Performance

In general, the effect of this checkin is positive on haswell. Haven't checked elsewhere yet.

### skylake-x gcc10.2, Relative to parent PR

It has an almost purely positive effect relative to the parent PR (the one where we handle empty objects separately):

| File                             | Blocks | master Cycles | PR Cycles | +Throughput | master Instr. | PR Instr. | -Instr. | master Branch Misses | PR Branch Misses | -Branch Misses | master Cache Misses | PR Cache Misses | -Cache Misses | master Cache Refs | PR Cache Refs | -Cache Refs |
| ---                              |    --: |   --: |   --: |   --: |   --: |   --: |   --: |    --: |    --: |   --: |    --: |    --: |   --: |    --: |    --: |   --: |
| citm_catalog                     |  26987 |  83.0 |  78.9 |    4% | 300.7 | 293.5 |    2% |   1827 |   1738 |    4% |   1639 |   1539 |    6% |  89439 |  89366 |    0% |
| canada                           |  35172 | 280.9 | 269.7 |    3% | 971.3 | 957.3 |    1% |  18841 |  18936 |    0% |  22517 |  28771 |  -27% | 154731 | 154773 |    0% |
| update-center                    |   8330 | 114.6 | 111.0 |    3% | 331.7 | 334.6 |    0% |   5994 |   5577 |    6% |     13 |     28 | -115% |  30980 |  31500 |   -1% |
| random                           |   7976 | 141.8 | 137.8 |    2% | 489.8 | 493.3 |    0% |   2036 |   1917 |    5% |      4 |     14 | -250% |  37029 |  34567 |    6% |
| apache_builds                    |   1988 |  85.7 |  83.4 |    2% | 302.2 | 303.9 |    0% |    107 |     97 |    9% |      5 |      0 |    0% |    131 |     86 |   34% |
| twitter                          |   9867 |  90.6 |  89.8 |    0% | 287.8 | 287.3 |    0% |   3392 |   3723 |   -9% |     10 |      0 |    0% |  32200 |  31006 |    3% |
| marine_ik                        |  46616 | 276.8 | 275.1 |    0% | 895.9 | 907.0 |   -1% |  53194 |  52758 |    0% | 104849 | 137813 |  -31% | 255984 | 255925 |    0% |
| mesh                             |  11306 | 290.8 | 289.3 |    0% | 977.8 | 994.3 |   -1% |  10051 |   8757 |   12% |    171 |    301 |  -76% |  60156 |  55079 |    8% |
| twitterescaped                   |   8787 | 184.4 | 183.7 |    0% | 491.4 | 494.4 |    0% |   4866 |   4872 |    0% |      1 |      1 |    0% |  27684 |  27962 |   -1% |
| numbers                          |   2345 | 239.4 | 238.5 |    0% | 790.8 | 803.6 |   -1% |   2193 |   2170 |    1% |      0 |      0 |    0% |      3 |     75 | -2400% |
| github_events                    |   1017 |  76.6 |  76.8 |    0% | 260.4 | 262.9 |    0% |     99 |    127 |  -28% |      2 |      0 |    0% |      2 |      0 |    0% |
| mesh.pretty                      |  24646 | 173.7 | 174.4 |    0% | 575.3 | 581.9 |   -1% |  13512 |  13292 |    1% |   1388 |   8533 | -514% |  87994 |  85411 |    2% |
| instruments                      |   3442 | 103.6 | 104.1 |    0% | 374.6 | 376.2 |    0% |    449 |    475 |   -5% |     12 |      6 |   50% |   1772 |   5097 | -187% |
| gsoc-2018                        |  51997 |  71.3 |  75.6 |   -6% | 164.0 | 164.9 |    0% |  30450 |  31266 |   -2% |  45546 |  74131 |  -62% | 168334 | 168489 |    0% |

### skylake-x gcc10.2, relative to master

It has a mixed (but generally positive) effect against master, when you include the mixed effect of the previous PR:

| File                             | Blocks | master Cycles | PR Cycles | +Throughput | master Instr. | PR Instr. | -Instr. | master Branch Misses | PR Branch Misses | -Branch Misses | master Cache Misses | PR Cache Misses | -Cache Misses | master Cache Refs | PR Cache Refs | -Cache Refs |
| ---                              |    --: |   --: |   --: |   --: |   --: |   --: |   --: |    --: |    --: |   --: |    --: |    --: |   --: |    --: |    --: |   --: |
| random                           |   7976 | 152.4 | 137.8 |    9% | 494.5 | 493.3 |    0% |   2069 |   1917 |    7% |     99 |     14 |   85% |  38598 |  34567 |   10% |
| update-center                    |   8330 | 122.6 | 111.0 |    9% | 343.5 | 334.6 |    2% |   5714 |   5577 |    2% |      5 |     28 | -459% |  30743 |  31500 |   -2% |
| apache_builds                    |   1988 |  90.5 |  83.4 |    7% | 309.6 | 303.9 |    1% |    109 |     97 |   11% |      0 |      0 |    0% |      0 |     86 |    0% |
| twitter                          |   9867 |  97.3 |  89.8 |    7% | 294.2 | 287.3 |    2% |   3571 |   3723 |   -4% |      0 |      0 |    0% |  33569 |  31006 |    7% |
| instruments                      |   3442 | 112.3 | 104.1 |    7% | 369.2 | 376.2 |   -1% |    473 |    475 |    0% |      5 |      6 |  -20% |    802 |   5097 | -535% |
| twitterescaped                   |   8787 | 197.5 | 183.7 |    6% | 512.6 | 494.4 |    3% |   5266 |   4872 |    7% |      5 |      1 |   80% |  29842 |  27962 |    6% |
| github_events                    |   1017 |  81.6 |  76.8 |    5% | 267.0 | 262.9 |    1% |    120 |    127 |   -5% |      0 |      0 |    0% |      0 |      0 |    0% |
| citm_catalog                     |  26987 |  83.6 |  78.9 |    5% | 307.0 | 293.5 |    4% |   1761 |   1738 |    1% |   1254 |   1539 |  -22% |  89395 |  89366 |    0% |
| canada                           |  35172 | 272.3 | 269.7 |    0% | 980.9 | 957.3 |    2% |  18825 |  18936 |    0% |  27293 |  28771 |   -5% | 154812 | 154773 |    0% |
| gsoc-2018                        |  51997 |  76.3 |  75.6 |    0% | 169.1 | 164.9 |    2% |  32178 |  31266 |    2% |  47011 |  74131 |  -57% | 168355 | 168489 |    0% |
| marine_ik                        |  46616 | 266.8 | 275.1 |   -3% | 903.9 | 907.0 |    0% |  52667 |  52758 |    0% | 107442 | 137813 |  -28% | 255891 | 255925 |    0% |
| numbers                          |   2345 | 231.0 | 238.5 |   -3% | 790.8 | 803.6 |   -1% |   2161 |   2170 |    0% |      0 |      0 |    0% |      0 |     75 |    0% |
| mesh.pretty                      |  24646 | 165.7 | 174.4 |   -5% | 579.5 | 581.9 |    0% |  13479 |  13292 |    1% |   1403 |   8533 | -508% |  88042 |  85411 |    2% |
| mesh                             |  11306 | 272.2 | 289.3 |   -6% | 986.9 | 994.3 |    0% |   9743 |   8757 |   10% |     59 |    301 | -410% |  60090 |  55079 |    8% |